requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
streamlit>=1.40.0
requests>=2.32.0
python-dotenv>=1.0.0
openai>=1.55.0
httpx>=0.27.0
python-telegram-bot>=21.0
